[cmd/opampsupervisor] Only report applying if config is changed (#39500)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Right now, the Supervisor will always report to the OpAMP server an
"Applying" status when it receives a remote config (assuming there's no
issue building the merged config). This is a problem because if the new
config is not different from the current config, the Supervisor never
restarts the Collector with the new config and the status "Applied" is
never reported to the server.

This change makes it so the Supervisor only reports the "Applying"
status to the OpAMP server if the new config is different from the
current and we intend to apply and restart the Collector.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes

<!--Describe what testing was performed and which tests were added.-->
#### Testing

<!--Describe the documentation added.-->
#### Documentation

<!--Please delete paragraphs that you did not use before submitting.-->