[processor/tailsampling] Drop policy type (#39668)
#### Description

This pull-request introduces a new policy type to explicitly drop traces
regardless of other policy decisions. The new `drop` policy type behaves
much like an `and`, however, it returns a `Dropped` decision when all of
its policies return `Sampled`. A `Dropped` decision takes precedence
over all others. The `Dropped` decision is not new, however, it was
unused and therefor safe to use for this purpose. This new policy type
is very **approachable** and capable.

This new policy type can be used to replace the top-level (not within an
`and` policy) use of `invert_match` to control the final sampling
decision. We could deprecate the current `invert_match` top-level
decision logic. The string, numeric, and boolean filter policies still
support `invert_match`, which continues to flip the decision for the
individual policy. Let `invert_match` be simple.

Example:

```
{
    name: drop-it-like-its-hot,
    type: drop,
    drop: {
        drop_sub_policy: 
        [
            {
                name: do-not-sample,
                type: boolean_attribute,
                boolean_attribute: { key: app.do_not_sample, value: true },
            }
        ]
    }
}
```

This is a reduced version of a previous PR
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/37760

#### Related Issues

-
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/36673
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/33656
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/36795
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34296
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/34085
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/29637
-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/27049
- Probably more 😅

---------

Signed-off-by: Sean Porter <portertech@gmail.com>