[chore][stanza/namedpipe] Refactor the namedpipe input to remove the fsnotify (#39529)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

The current logic is that the input starts 2 goroutines, one for
watching for write events, and another one listens from `watcher.C` and
kick off the process. In other words, it relies on a write event to
start to read from the pipe.

When an active writer is writing to the pipe, and the collector exits
for whatever reason (or simply can't consume fast enough), the pipe
buffer can be filled and no write can come in. When no write happens,
the watcher doesn't receive any write event, and won't kick off the
process to read the pipe. This is effectively a deadlock, the writer
can't write to the pipe because the pipe buffer is full, and the
collector won't read from the pipe because no write event triggered.

Refactor the namedpipe input to get rid of the watcher. The input will
now keep reading from pipe. If no data or no external writer, it blocks
at the `scan.Scan()` wait for new data or an error when pipe is closed.

#### How to reproduce the bug
A simple python script to write to a named pipe periodically:

```python
import os
import time

PIPE_PATH = "/tmp/my_named_pipe"

try:
    fd = os.open(PIPE_PATH, os.O_RDWR | os.O_APPEND | os.O_NONBLOCK)
except OSError as e:
    print(f"Failed to open pipe: {e}")
    exit(1)

msg = b"x" * 1024 * 10 + b"\n" # 10 KB message with a newline at the end

while True:
    try:
        os.write(fd, msg)
        print("write ok")
    except BlockingIOError:
        print("PIPE FULL - write blocked")
    time.sleep(1)
```

The minimal Otel collector config:
```yaml
receivers:
  namedpipe:
    path: /tmp/my_named_pipe
    mode: 0600

exporters:
  debug:
    verbosity: basic

service:
  pipelines:
    logs:
      receivers: [namedpipe]
      processors: []
      exporters: [debug]
```

1. Start the python script to write the pipe to full (when seeing "PIPE
FULL - write blocked"). It's usually the 7th write if the pipe buffer is
64K (default value).
2. Start the otel collector and see it doesn't process anything from the
pipe.

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Passed all current unit tests. Also tested locally with the steps above
to confirm it reads from the pipe from the start.

---------

Signed-off-by: Mengnan Gong <namco1992@gmail.com>
Co-authored-by: Daniel Jaglowski <jaglows3@gmail.com>