[tailsamplingprocessor] Fix the misleading decision timer latency metric (#39761)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

It will now measure the total time spent on one tick of the decision
timer. This is more useful, since the decision tick loop needs to run at
least once per second for the tsp to keep up with incoming traces.



<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #38502