[receiver/googlecloudpubsub] add ignore_encoding_error configuration (#39839)
#### Description
Introduce a setting to ignore errors when the configured encoder. It's
advised to set this to `true` when using
a custom encoder, and use `receiver.googlecloudpubsub.encoding_error`
metric to monitor the number of errors.
Ignoring the error will cause the receiver to drop the message.

#### Link to tracking issue
#38164

#### Testing
Tested with a custom encoder and introducing bogus message to the topic.

#### Documentation
Added the configuration setting to the README