[receiver/prometheus] Upgrade receiver.prometheus.removeLegacyResourceAttributes feature flag (#39803)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
This feature flag was introduced in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/32829
to resolve #32814. It was first released in `v0.103.0`, on June 19, 2024
which was 10 months ago. The feature gate is to remove deprecated
resource attributes in favor of the semantic convention equivalents.
This is a breaking change and the feature flag will now be enabled by
default.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue

https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/32814

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Updated testing to account for the feature flag being enabled by
default.