[pkg/winperfcounters] Add support to retrieve raw values (#39835)
Adding public methods that allow to retrieve the raw values of Windows
performance counters. This is needed to allow removal of the vendored
code under `receiver/hostmetricsreceiver/internal/perfcounters/` used
only by that component. This vendored code retrieves raw values of
counters instead of calculated ones, e.g.: it doesn't retrieve rate, but
the actual values used to calculate the rate of a counter. See
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/38858#issuecomment-2748552995
for details.